Create MultiplicationTable
Using a for loop: Write a program that takes an integer input from the user and prints out a
multiplication table for that number from 1 to 10. The program should use a for loop to iterate
through the numbers from 1 to 10 and print out the product of the input number and the
current iteration number. For example, if the user inputs 5, the program should output:
5 x 1 = 5 5 x 2 = 10 5 x 3 = 15 ... 5 x 10 = 50